:sparkles: display grade averages on the grade view (close #12)
we use a static json file to map from course of study short names to long names. Each long name corresponds to one SPO pdf file on thi.de which is automatically downloaded and parsed using the SPO parser. While it works reasonably well it might miss some grade weights because of the nature of the pdf->weight conversion

also the short name -> long name mapping has to be created manually and is currently incomplete